changed hashing of Scene Creation to better capture involved subgraphs

+++ algorithms/algorithms.go
@@ -11,7 +11,7 @@ type Algorithm interface {
 
 type UpdateAlgorithm interface {
 	Name() string
-	FindDecompUpdate(currentGraph Graph, savedScenes map[uint32]SceneValue, savedCache Cache) Decomp
+	FindDecompUpdate(currentGraph Graph, savedScenes map[uint64]SceneValue, savedCache Cache) Decomp
 	SetWidth(K int)
 	GetCache() Cache
 }
